src/requirements.txt
sanic==0.7.0
aiocache==0.8.0
von-agent==0.5.11
python3-indy==1.3.1-dev-408
jsonschema>=2.6.0
